+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.16.4
+
+* added embedding configuration documentation in schema JSON
+    * this functionality does not require updating satella