+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * added embedding configuration documentation in schema JSON
     * this functionality does not require updating satella
+* added `strip_afterwards` to config schema file_contents