+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.16.5
+
+* added `reraise_as`