+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-# v2.16.5
-
-* added `reraise_as`
+# v2.16.6